requirements.txt
PyYAML>=6.0.1
SQLAlchemy>=2.0.30
requests>=2.31.0
Jinja2>=3.1.4
python-dotenv>=1.0.1
tenacity>=8.3.0
dataclasses-json>=0.6.4
typer>=0.12.3
beautifulsoup4>=4.12.3
httpx>=0.27.2
fake-useragent>=1.5.1
openai>=1.51.2
fastapi>=0.115.0
uvicorn>=0.30.6
sse-starlette>=2.1.0
